Migrate metrics to `metric.go`
(cherry picked from commit ce005fc6b4b1aece022a55445da644d7b0d87040)
